requirements.txt
aiogram==3.4.1
pydantic-settings==2.2.1
sqlalchemy==2.0.25
aiosqlite==0.19.0
asyncpg==0.29.0
anthropic>=0.40.0
python-docx==1.1.0
python-dotenv==1.0.1
